* **BREAKING** after a longer research i now settled on size_t as the index and capacity for the datastructures. So from uintptr_t to unsigned int now to size_t. I really don't want my own unsigned <whatevr> define in here and also not NSUInteger which must be uintptr_t

+++ obsolete/uniquepointer/mulle--uniquepointerarray.c
@@ -24,7 +24,7 @@ static int  _pointer_compare( void *p_a, void *p_b, void *unused)
 
 
 MULLE_C_NONNULL_FIRST_THIRD
-void   *mulle_pointers_bsearch( void **buf, unsigned int n, void *search)
+void   *mulle_pointers_bsearch( void **buf, size_t n, void *search)
 {
    long   first;
    long   last;
